requirements.txt
flask
loguru
pycryptodome